Process ACCEPT response in overlay service
