+++ newsfragments/345.changed.md
@@ -1,0 +1,1 @@
+Enable the gossiping of content to peers, by changing how we process `ACCEPT` messages.
